#58 ryft doesn't support file moving
in order to keep INDEX or DATA files we have
to pass provided names to ryftprim directly

+++ search/ryftprim/ryftprim.go
@@ -89,22 +89,37 @@ func (engine *Engine) prepare(task *Task, cfg *search.Config) error {
 
 	// INDEX results file
 	if len(task.IndexFileName) != 0 {
-		// file path relative to `ryftone` mountpoint (including just instance)
-		file := filepath.Join(engine.Instance, task.IndexFileName)
-		args = append(args, "-oi", file)
+		if len(cfg.KeepIndexAs) != 0 {
+			task.IndexFileName = cfg.KeepIndexAs
+			task.KeepIndexFile = true
+			if !strings.HasSuffix(task.IndexFileName, ".txt") {
+				// ryft adds .txt anyway, so if this extension is missed
+				// other code won't properly work!
+				task.IndexFileName += ".txt"
+				task.log().WithField("index", task.IndexFileName).
+					Warnf("[%s]: index file name was updated to have TXT extension", TAG)
+			}
+		} else {
+			// file path relative to `ryftone` mountpoint (including just instance)
+			task.IndexFileName = filepath.Join(engine.Instance, task.IndexFileName)
+		}
+		args = append(args, "-oi", task.IndexFileName)
 	}
 
 	// DATA results file
 	if len(task.DataFileName) != 0 {
-		// file path relative to `ryftone` mountpoint (including just instance)
-		file := filepath.Join(engine.Instance, task.DataFileName)
-		args = append(args, "-od", file)
+		if len(cfg.KeepDataAs) != 0 {
+			task.DataFileName = cfg.KeepDataAs
+			task.KeepDataFile = true
+		} else {
+			// file path relative to `ryftone` mountpoint (including just instance)
+			task.DataFileName = filepath.Join(engine.Instance, task.DataFileName)
+		}
+		args = append(args, "-od", task.DataFileName)
 	}
 
 	// assign command line
 	task.tool_args = args
-	task.KeepIndexFileAs = cfg.KeepIndexAs
-	task.KeepDataFileAs = cfg.KeepDataAs
 
 	return nil // OK
 }
@@ -268,31 +283,13 @@ func (engine *Engine) finish(err error, task *Task, res *search.Result) {
 	}
 
 	// cleanup: remove INDEX&DATA files at the end of processing
-	if len(task.KeepIndexFileAs) != 0 {
-		task.log().WithField("from", filepath.Join(engine.Instance, task.IndexFileName)).
-			WithField("to", task.KeepIndexFileAs).Debugf("[%s]: moving INDEX file", TAG)
-		if err := engine.moveFile(task.IndexFileName, task.KeepIndexFileAs); err != nil {
-			task.log().WithError(err).WithField("file", task.KeepIndexFileAs).Warnf("[%s]: failed to move INDEX file", TAG)
-			res.ReportError(fmt.Errorf("failed to move INDEX file to %q: %s", task.KeepIndexFileAs, err))
-		} else {
-			task.IndexFileName = "" // prevent to remove later
-		}
-	}
-	if len(task.KeepDataFileAs) != 0 {
-		task.log().WithField("from", filepath.Join(engine.Instance, task.DataFileName)).
-			WithField("to", task.KeepDataFileAs).Debugf("[%s]: moving DATA file", TAG)
-		if err := engine.moveFile(task.DataFileName, task.KeepDataFileAs); err != nil {
-			task.log().WithError(err).WithField("file", task.KeepDataFileAs).Warnf("[%s]: failed to move DATA file", TAG)
-			res.ReportError(fmt.Errorf("failed to move DATA file to %q: %s", task.KeepDataFileAs, err))
-		} else {
-			task.DataFileName = "" // prevent to remove later
-		}
-	}
-	if !engine.KeepResultFiles {
+	if !engine.KeepResultFiles && !task.KeepIndexFile {
 		if err := engine.removeFile(task.IndexFileName); err != nil {
 			task.log().WithError(err).Warnf("[%s]: failed to remove INDEX file", TAG)
 			// WARN: error actually ignored!
 		}
+	}
+	if !engine.KeepResultFiles && !task.KeepDataFile {
 		if err := engine.removeFile(task.DataFileName); err != nil {
 			task.log().WithError(err).Warnf("[%s]: failed to remove DATA file", TAG)
 			// WARN: error actually ignored!
@@ -309,7 +306,7 @@ func (engine *Engine) processIndex(task *Task, res *search.Result) {
 	task.log().Debugf("[%s]: start INDEX processing...", TAG)
 
 	// try to open INDEX file: if operation is cancelled `file` is nil
-	path := filepath.Join(engine.MountPoint, engine.Instance, task.IndexFileName)
+	path := filepath.Join(engine.MountPoint, task.IndexFileName)
 	file, err := task.openFile(path, engine.OpenFilePollTimeout, task.cancelIndexChan)
 	if err != nil {
 		task.log().WithError(err).WithField("path", path).
@@ -381,7 +378,7 @@ func (engine *Engine) processData(task *Task, res *search.Result) {
 	task.log().Debugf("[%s]: start DATA processing...", TAG)
 
 	// try to open DATA file: if operation is cancelled `file` is nil
-	path := filepath.Join(engine.MountPoint, engine.Instance, task.DataFileName)
+	path := filepath.Join(engine.MountPoint, task.DataFileName)
 	file, err := task.openFile(path, engine.OpenFilePollTimeout, task.cancelDataChan)
 	if err != nil {
 		task.log().WithError(err).WithField("path", path).
@@ -444,24 +441,8 @@ func (engine *Engine) prepareQuery(query string) string {
 // removeFile removes INDEX or DATA file.
 func (engine *Engine) removeFile(name string) error {
 	if len(name) != 0 {
-		path := filepath.Join(engine.MountPoint,
-			engine.Instance, name) // full path
+		path := filepath.Join(engine.MountPoint, name) // full path
 		err := os.RemoveAll(path)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil // OK
-}
-
-// moveFile moves INDEX or DATA file to a new location.
-func (engine *Engine) moveFile(from string, to string) error {
-	if len(from) != 0 {
-		fromPath := filepath.Join(engine.MountPoint,
-			engine.Instance, from) // full path
-		toPath := filepath.Join(engine.MountPoint, to)
-		err := os.Rename(fromPath, toPath)
 		if err != nil {
 			return err
 		}
